rename leaderboardscoring to leaderboardscoringapp
