+++ cmd/leaderboardscoring/command/migrate.go
@@ -1,7 +1,7 @@
 package command
 
 import (
-	"github.com/gocasters/rankr/leaderboardscoring"
+	"github.com/gocasters/rankr/leaderboardscoringapp"
 	"github.com/gocasters/rankr/pkg/config"
 	"github.com/gocasters/rankr/pkg/migrator"
 	"github.com/spf13/cobra"
@@ -29,7 +29,7 @@ func init() {
 }
 
 func migrate() {
-	var cfg leaderboardscoring.Config
+	var cfg leaderboardscoringapp.Config
 
 	workingDir, err := os.Getwd()
 	if err != nil {
